optimize for timesteps

+++ cldm/ddim_hacked.py
@@ -150,7 +150,7 @@ class DDIMSampler(object):
 
         for i, step in enumerate(iterator):
             index = total_steps - i - 1
-            ts = torch.full((b,), step, device=device, dtype=torch.long)
+            ts = torch.full((b,), step, device=device, dtype=torch.int32)
 
             if mask is not None:
                 assert x0 is not None
